新建文件/文件夹

+++ app/controllers/directories_controller.rb
@@ -18,6 +18,15 @@ class DirectoriesController < ApplicationController
   def edit
   end
 
+  def create
+    code = params[:filetype]=='file' ? File.open(params[:path], 'w+') : Dir.mkdir(params[:path], 0700)
+    if (params[:filetype]=='file' && code) || (params[:filetype]=='dir' && code==0)
+      render json: {msg: 'ok'}
+    else
+      render json: {msg: 'error'}, status: 400
+    end
+  end
+
   def update
     file_path = "#{params[:root_path]}/#{params[:extra_path]}"
     File.open(file_path, 'w+') do |file|
@@ -36,8 +45,10 @@ class DirectoriesController < ApplicationController
   end
 
   def destroy
-    num = File.delete(params[:path])
-    if num>0
+    is_file = File.file?(params[:path])
+    num = File.delete(params[:path]) if is_file
+    code = Dir.delete(params[:path]) unless is_file
+    if (is_file && num>0) || (!is_file && code==0)
       render json: {msg: 'ok'}
     else
       render json: {msg: 'error'}, status: 400
